DeckQueryRepo- 덱 추천 관련해서 querydsl 코드 수정하ㅇ였습니다!

+++ src/main/java/com/example/thirdtool/Deck/domain/repository/DeckQueryRepository.java
@@ -29,8 +29,8 @@ public class DeckQueryRepository {
                         deck.id,
                         deck.name,
                         lp.score.avg(),
-                        lp.badCount.sum(),
-                        lp.badCount.add(lp.normalCount).add(lp.goodCount).add(lp.greatCount).sum(),
+                        lp.badCount.sum().longValue(),
+                        lp.badCount.add(lp.normalCount).add(lp.goodCount).add(lp.greatCount).sum().longValue(),
                         deck.lastAccessed.max(),
                         card.id.count()
                                                ))
